@@ -49,74 +49,70 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~10.0.9",
-    "@angular/cdk": "^10.1.3",
-    "@angular/common": "~10.0.9",
-    "@angular/compiler": "~10.0.9",
-    "@angular/core": "~10.0.9",
-    "@angular/elements": "^10.0.12",
-    "@angular/forms": "~10.0.9",
-    "@angular/material": "^10.1.3",
-    "@angular/platform-browser": "~10.0.9",
-    "@angular/platform-browser-dynamic": "~10.0.9",
-    "@angular/router": "~10.0.9",
-    "@senzing/rest-api-client-ng": "^2.3.0",
-    "@senzing/sdk-graph-components": "^2.4.1",
+    "@angular/animations": "~13.0.3",
+    "@angular/cdk": "^13.0.3",
+    "@angular/common": "~13.0.3",
+    "@angular/compiler": "~13.0.3",
+    "@angular/core": "~13.0.3",
+    "@angular/elements": "^13.0.3",
+    "@angular/forms": "~13.0.3",
+    "@angular/material": "^13.0.3",
+    "@angular/platform-browser": "~13.0.3",
+    "@angular/platform-browser-dynamic": "~13.0.3",
+    "@angular/router": "~13.0.3",
+    "@senzing/rest-api-client-ng": "~3.0.0",
+    "@senzing/sdk-graph-components": "~3.0.0",
+    "acorn": "^8.6.0",
+    "ajv": "^6.12.6",
     "d3": "^5.9.1",
-    "document-register-element": "^1.14.5",
+    "document-register-element": "^1.14.10",
     "html2canvas": "^1.0.0-rc.7",
-    "jspdf": "^2.3.1",
-    "ngx-json-viewer": "^2.4.0",
-    "rxjs": "~6.5.5",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.10.3"
+    "jspdf": "^2.4.0",
+    "ngx-json-viewer": "^3.0.2",
+    "rxjs": "~7.4.0",
+    "sass": "^1.44.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1002.0",
-    "@angular-devkit/build-ng-packagr": "~0.1000.7",
-    "@angular-devkit/core": "^10.2.0",
-    "@angular-devkit/schematics": "^10.2.0",
-    "@angular/cli": "~10.0.6",
-    "@angular/compiler-cli": "~10.0.9",
-    "@angular/language-service": "^10.0.14",
+    "@angular-devkit/build-angular": "^13.0.4",
+    "@angular-devkit/core": "^13.0.4",
+    "@angular-devkit/schematics": "^13.0.4",
+    "@angular/cli": "~13.0.4",
+    "@angular/compiler-cli": "~13.0.3",
+    "@angular/language-service": "^13.0.3",
     "@compodoc/compodoc": "^1.1.11",
     "@types/d3": "^5.7.1",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/jspdf": "^1.3.0",
     "@types/node": "^12.11.1",
-    "angular-in-memory-web-api": "^0.11.0",
-    "cheerio": "^1.0.0-rc.3",
-    "codelyzer": "^6.0.0",
+    "angular-in-memory-web-api": "^0.13.0",
+    "cheerio": "^1.0.0-rc.10",
     "concat": "^1.0.3",
-    "concurrently": "^5.3.0",
+    "concurrently": "^6.4.0",
     "fs-extra": "^9.0.1",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
-    "karma-brief-reporter": "^0.2.1",
+    "jasmine-core": "~3.10.0",
+    "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~3.3.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "karma-mocha-reporter": "^2.2.5",
-    "karma-summary-reporter": "^1.8.0",
-    "ng-packagr": "^10.1.1",
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "karma-summary-reporter": "~3.0.0",
+    "ng-packagr": "^13.1.1",
     "ngx-webstorage-service": "^4.1.0",
-    "node-sass": "^4.14.1",
     "protractor": "~7.0.0",
     "rimraf": "^3.0.2",
     "sass-burger": "^1.3.1",
-    "schematics-utilities": "^2.0.2",
     "scss-bundle": "^3.1.2",
     "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~3.9.5",
+    "typescript": "~4.4.4",
     "wait-on": "^5.2.1"
   },
   "optionalDependencies": {
+    "codelyzer": "^6.0.0",
     "express": "^4.17.1",
-    "http-proxy-middleware": "^1.0.6",
+    "http-proxy-middleware": "^2.0.1",
     "serve-static": "^1.14.1"
   }
 }
