@@ -1,6 +1,6 @@
 {
   "name": "@senzing/sdk-components-ng",
-  "version": "2.2.6",
+  "version": "3.0.0",
   "private": false,
   "keywords" :["Angular","Library","Senzing"],
   "license" : "SEE LICENSE IN LICENSE",
@@ -10,21 +10,20 @@
   },
   "homepage" :"https://senzing.github.io/sdk-components-ng/",
   "dependencies": {
-    "schematics-utilities": "^2.0.2",
-    "@senzing/rest-api-client-ng": "^2.3.0",
-    "d3": "^5.9.1",
-    "jspdf": "^2.3.1",
-    "html2canvas": "^1.0.0-rc.3",
-    "ngx-json-viewer": "^2.4.0",
     "@types/d3": "^5.7.1",
-    "@types/jspdf": "^1.3.0"
+    "@types/jspdf": "^1.3.0",
+    "d3": "^5.9.1",
+    "html2canvas": "^1.0.0-rc.3",
+    "jspdf": "^2.3.1",
+    "ngx-json-viewer": "^2.4.0",
+    "tslib": "^2.3.0"
   },
   "peerDependencies": {
-    "@angular/cdk": "^10.1.3",
-    "@angular/common": "~10.0.9",
-    "@angular/core": "~10.0.9",
-    "@angular/material": "^10.1.3",
-    "@senzing/sdk-graph-components": "^2.4.1"
-  },
-  "schematics": "./schematics/collection.json"
+    "@angular/cdk": "^13.0.3",
+    "@angular/common": "~13.0.3",
+    "@angular/core": "~13.0.3",
+    "@angular/material": "~13.0.3",
+    "@senzing/rest-api-client-ng": "^3.0.0",
+    "@senzing/sdk-graph-components": "^3.0.0"
+  }
 }
