@@ -1,9 +1,8 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
-import { CommonModule, TitleCasePipe } from '@angular/common';
-import { HttpClientModule, HttpClient } from '@angular/common/http';
-import { SenzingSdkModule, SzRestConfiguration, SzPoweredByComponent  } from '@senzing/sdk-components-ng';
+import { HttpClientModule } from '@angular/common/http';
+import { SenzingSdkModule, SzRestConfiguration  } from '@senzing/sdk-components-ng';
 import { SenzingSdkGraphModule } from '@senzing/sdk-graph-components';
 import { AppComponent } from './app.component';
 
