@@ -136,9 +136,9 @@ export class SzEntityDetailHeaderComponent implements OnInit, OnDestroy {
   private getGenderFromFeatures(features: {[key: string] : SzEntityFeature[]} | undefined | null): string | undefined {
     if(features){
       //console.warn('getGenderFromFeatures: ', features.GENDER);
-      if(features.GENDER){
+      if(features['GENDER']){
         // has gender
-        let _gender = features.GENDER;
+        let _gender = features['GENDER'];
         if(_gender.some) {
           let _female = _gender.some( (val: {primaryValue: string, usageType: any, duplicateValues: any} ) => {
             return val.primaryValue === "F";
@@ -151,6 +151,7 @@ export class SzEntityDetailHeaderComponent implements OnInit, OnDestroy {
         }
       }
     }
+    return undefined;
   }
   /**
    * returns true if a Entities features collection can identify it as a person.
@@ -166,11 +167,12 @@ export class SzEntityDetailHeaderComponent implements OnInit, OnDestroy {
           hasPersonKey = true; // has key
         }
       });
-      if(features && (features.ORG || features.NAME_ORG)){
+      if(features && (features['ORG'] || features['NAME_ORG'])){
         hasBusinessKey = true;
       }
       return (!hasBusinessKey && hasPersonKey);
     }
+    return false;
   }
   /**
    * returns the svg view box to use for the primary icon
@@ -244,7 +246,7 @@ export class SzEntityDetailHeaderComponent implements OnInit, OnDestroy {
     } else if(maxWidth){
       return (`(max-width: ${maxWidth}px)`);
     }
-    return
+    return undefined;
   }
 
   ngOnInit() {
