@@ -479,7 +479,7 @@ export class SzSearchByIdComponent implements OnInit, OnDestroy {
     } else if(maxWidth){
       return (`(max-width: ${maxWidth}px)`);
     }
-    return
+    return undefined;
   }
 
   /**
